content/html/document/test/test_non-ascii-cookie.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=784367
-->
<head>
  <meta charset="utf-8">
  <title>Test for non-ASCII document.cookie</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=784367">Mozilla Bug 784367</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for non-ASCII document.cookie **/
var c = document.cookie;
is(document.cookie, 'abc=012©ABC\ufffdDEF', "document.cookie should be decoded as UTF-8");
var newCookie = 'def=∼≩≭≧∯≳≲≣∽≸≸∺≸∠≯≮≥≲≲≯≲∽≡≬≥≲≴∨∱∩∾';
document.cookie = newCookie;
is(document.cookie, c + '; ' + newCookie, "document.cookie should be encoded as UTF-8");
var date1 = new Date();
date1.setTime(0);
document.cookie = newCookie + 'def=;expires=' + date1.toGMTString();
</script>
</pre>
</body>
</html>
